Refactor to modularize task management logic
Split `task.js` into dedicated modules for cleaner structure. This modularisation improves maintainability and supports scalability. Updated `index.html` to reflect the new entry point, `main.js`.
